refactor: add validation middleware
