+++ backend/route/contactRouter.js
@@ -3,23 +3,9 @@ const { getContacts, getContactDetail } = require('../service');
 
 const contactRouter = express.Router();
 
-contactRouter.get('/contacts', async (req, res) => {
+contactRouter.get('/', async (req, res) => {
   try {
-    const options = {
-      pagination: {
-        page: parseInt(req.query.page, 10) || 0,
-        rowsPerPage: parseInt(req.query.rowsPerPage, 10) || 10,
-      },
-      sort: {
-        key: req.query.sortKey || 'UserID',
-        order: req.query.sortOrder || 'ASC',
-      },
-      filter: {
-        key: req.query.filterKey || '',
-        value: req.query.filterValue || '',
-      },
-    };
-    const result = await getContacts(options);
+    const result = await getContacts(req.body.options);
     const response = {
       status: 'success',
       contacts: result.contacts,
@@ -35,7 +21,7 @@ contactRouter.get('/contacts', async (req, res) => {
   }
 });
 
-contactRouter.get('/contacts/:userID', async (req, res) => {
+contactRouter.get('/:userID', async (req, res) => {
   try {
     const { userID } = req.params;
     const detail = await getContactDetail(userID);
